Moved class to tests
